refactor: 유저 저장 로직 최적화
