+++ team1-api/src/main/java/goorm/deepdive/team1/api/Team1ApiApplication.java
@@ -5,9 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class Team1ApiApplication {
-
 	public static void main(String[] args) {
 		SpringApplication.run(Team1ApiApplication.class, args);
 	}
-
 }
