+++ team1-domain/src/main/java/goorm/deepdive/team1/domain/address/exception/KakaoApiException/AddressNotFoundException.java
@@ -1,8 +1,8 @@
-package goorm.deepdive.team1.common.exception.KakaoApiException;
+package goorm.deepdive.team1.domain.address.exception.KakaoApiException;
 
 import goorm.deepdive.team1.common.exception.CustomException;
 
-import static goorm.deepdive.team1.common.exception.KakaoApiException.KakaoApiExceptionCode.ADDRESS_NOT_FOUND;
+import static goorm.deepdive.team1.domain.address.exception.KakaoApiException.KakaoApiExceptionCode.ADDRESS_NOT_FOUND;
 
 public class AddressNotFoundException extends CustomException {
     public AddressNotFoundException() {
