+++ team1-domain/src/main/java/goorm/deepdive/team1/domain/address/exception/KakaoApiException/UnauthorizedException.java
@@ -1,8 +1,8 @@
-package goorm.deepdive.team1.common.exception.KakaoApiException;
+package goorm.deepdive.team1.domain.address.exception.KakaoApiException;
 
 import goorm.deepdive.team1.common.exception.CustomException;
 
-import static goorm.deepdive.team1.common.exception.KakaoApiException.KakaoApiExceptionCode.UNAUTHORIZED;
+import static goorm.deepdive.team1.domain.address.exception.KakaoApiException.KakaoApiExceptionCode.UNAUTHORIZED;
 
 public class UnauthorizedException extends CustomException {
     public UnauthorizedException() {
